+++ app/src/main/java/ru/meseen/dev/androidacademy/data/retrofit/pojo/items/MovieAdditionalDataResponse.kt
@@ -42,7 +42,7 @@ data class MovieAdditionalDataResponse(
     val backdropPath: String? = null,
 
     @SerialName("revenue")
-    val revenue: Int = -1,
+    val revenue: Long = -1,
 
     @SerialName("genres")
     val genres: List<GenresItem> = listOf(),
